+++ Demo Projects/Simple Track Playback/Simple Track Playback/ViewController.m
@@ -14,6 +14,7 @@
 @property (weak, nonatomic) IBOutlet UILabel *albumLabel;
 @property (weak, nonatomic) IBOutlet UILabel *artistLabel;
 @property (weak, nonatomic) IBOutlet UIImageView *coverView;
+@property (weak, nonatomic) IBOutlet UIActivityIndicatorView *spinner;
 
 @property (nonatomic, strong) SPTTrackPlayer *trackPlayer;
 
@@ -65,9 +66,51 @@
 		SPTAlbum *album = (SPTAlbum *)self.trackPlayer.currentProvider;
 		self.titleLabel.text = [album.tracks[index] name];
 		self.albumLabel.text = album.name;
-		self.artistLabel.text = album.artist.name;
-		self.coverView.image = [UIImage imageNamed:@"coverart"];
+		self.artistLabel.text = ((SPTArtist *)album.artists.firstObject).name;
+		self.coverView.image = nil;
+
+		[self loadCoverArt];
 	}
+}
+
+-(void)loadCoverArt {
+
+	if (![self.trackPlayer.currentProvider isKindOfClass:[SPTAlbum class]]) {
+		NSLog(@"-loadCoverArt makes the assumption that the currently playing provider is an album. Sorry!");
+		return;
+	}
+
+	SPTAlbum *album = (SPTAlbum *)self.trackPlayer.currentProvider;
+	NSURL *imageURL = album.largestCover.imageURL;
+
+	if (imageURL == nil) {
+		NSLog(@"Album %@ doesn't have any images!", album);
+		self.coverView.image = nil;
+		return;
+	}
+
+	[self.spinner startAnimating];
+
+	// Pop over to a background queue to load the image over the network.
+	dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
+
+		NSError *error = nil;
+		UIImage *image = nil;
+		NSData *imageData = [NSData dataWithContentsOfURL:imageURL options:0 error:&error];
+
+		if (imageData != nil) {
+			image = [UIImage imageWithData:imageData];
+		}
+
+		// …and back to the main queue to display the image.
+		dispatch_async(dispatch_get_main_queue(), ^{
+			[self.spinner stopAnimating];
+			self.coverView.image = image;
+			if (image == nil) {
+				NSLog(@"Couldn't load cover image with error: %@", error);
+			}
+		});
+	});
 }
 
 -(void)handleNewSession:(SPTSession *)session {