+++ Spotify.framework/Versions/A/Headers/SPTTrack.h
@@ -42,6 +42,12 @@
 /** The Spotify URI of the track. */
 @property (nonatomic, readonly, copy) NSURL *uri;
 
+/** The HTTP open.spotify.com URL of the track. */
+@property (nonatomic, readonly, copy) NSURL *sharingURL;
+
+/** The HTTP URL of a 30-second preview MP3 of the track. */
+@property (nonatomic, readonly, copy) NSURL *previewURL;
+
 /** The duration of the track. */
 @property (nonatomic, readonly) NSTimeInterval duration;
 
@@ -54,14 +60,17 @@
 /** The track number of the track. I.e., if it's the first track on the album this will be `1`. */
 @property (nonatomic, readonly) NSInteger trackNumber;
 
-/** The popularity of the track as a value between 0.0 (least popular) to 1.0 (most popular). */
+/** The disc number of the track. I.e., if it's the first disc on the album this will be `1`. */
+@property (nonatomic, readonly) NSInteger discNumber;
+
+/** The popularity of the track as a value between 0.0 (least popular) to 100.0 (most popular). */
 @property (nonatomic, readonly) double popularity;
+
+/** Returns `YES` if the track is flagged as explicit, otherwise `NO`. */
+@property (nonatomic, readonly) BOOL flaggedExplicit;
 
 /** Any external IDs of the track, such as the ISRC code. */
 @property (nonatomic, readonly, copy) NSDictionary *externalIds;
-
-/** `YES` if the track is available for playback, otherwise `NO`. */
-@property (nonatomic, readonly) BOOL availableForPlayback;
 
 /** An array of ISO 3166 country codes in which the album is available. */
 @property (nonatomic, readonly, copy) NSArray *availableTerritories;